examples/hubspot/requirements.txt
slack-sdk
backoff
python-dotenv
duckyai
fastapi
uvicorn[standard]
groq
slack_bolt
requests
hubspot-api-client